Refactor: api 파일 정리
