--- src/api/createCard.ts
@@ -1,10 +1,0 @@
-import { axiosInstance } from "./axiosInstance";
-
-export default async function createCard(url: string, folderId: string) {
-  const response = await axiosInstance.post(`/links`, {
-    url: url,
-    folderId: folderId,
-  });
-
-  return response;
-}
