--- src/api/deleteFolder.ts
@@ -1,6 +1,0 @@
-import { axiosInstance } from "./axiosInstance";
-
-export default async function deleteFolder(folderId: string) {
-  const response = await axiosInstance.delete(`/folders/${folderId}`);
-  return response;
-}
