--- src/api/editFolder.ts
@@ -1,9 +1,0 @@
-import { axiosInstance } from "./axiosInstance";
-
-export default async function editFolder(folderId: string, name: string) {
-  const response = await axiosInstance.put(`/folders/${folderId}`, {
-    name: name,
-  });
-
-  return response;
-}
