--- src/api/getFolderInfo.tsx
@@ -1,6 +1,0 @@
-import { axiosInstance } from "./axiosInstance";
-
-export default async function getFolderInfo(folderId: string) {
-  const response = await axiosInstance.get(`/folders/${folderId}`);
-  return response.data;
-}
