--- src/api/getSignin.tsx
@@ -1,9 +1,0 @@
-import { axiosInstance } from "./axiosInstance";
-
-export default async function getSignIn(email = "", password = "") {
-  const response = await axiosInstance.post("auth/sign-in", {
-    email: email,
-    password: password,
-  });
-  return response;
-}
