+++ src/api/getUser.tsx
@@ -1,7 +1,7 @@
 import { UserType } from "@/types/UserType";
 import { axiosInstance } from "./axiosInstance";
 
-export default async function getUser() {
+export async function getUser() {
   const response = await axiosInstance.get<[UserType]>(`/users`);
   return response.data[0];
 }
