+++ src/modals/LinkAddModal/LinkAddModal.tsx
@@ -1,13 +1,13 @@
+import { useRouter } from "next/router";
 import toast from "react-hot-toast";
 import { useMutation, useQueryClient, useQuery } from "@tanstack/react-query";
 import { useForm } from "react-hook-form";
 
-import getFolderList from "@/api/getFolderList";
-import createCard from "@/api/createCard";
+import { getFolderList } from "@/api/getFolderCRUDApi";
+import { createCard } from "@/api/getCardCRUDApi";
 import ModalCreator from "@/modals/ModalCreator";
 
 import styles from "./LinkAddModal.module.scss";
-import { useRouter } from "next/router";
 
 interface ModalProps {
   link: string;
