+++ src/modals/LinkDeleteModal/LinkDeleteModal.tsx
@@ -1,9 +1,11 @@
-import { useMutation, useQueryClient } from "@tanstack/react-query";
-import ModalCreator from "../ModalCreator";
-import styles from "./LinkDeleteModal.module.scss";
-import deleteCard from "@/api/deleteCard";
-import toast from "react-hot-toast";
 import { useRouter } from "next/router";
+import toast from "react-hot-toast";
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+
+import ModalCreator from "../ModalCreator";
+import { deleteCard } from "@/api/getCardCRUDApi";
+
+import styles from "./LinkDeleteModal.module.scss";
 
 interface LinkDeleteModalProps {
   cardId: string;
